requirements.txt
langgraph>=0.2.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
duckduckgo-search>=6.3.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
pydantic>=2.0.0
streamlit>=1.31.0
reportlab>=4.0.0
pypdf>=4.0.0
agentops>=0.3.0
langchain-openai>=0.1.0
langsmith>=0.1.0
tenacity>=8.2.0
certifi>=2024.0.0
